move package/versions to manager/npm
