+++ lib/manager/docker/package.js
@@ -1,6 +1,6 @@
 const semver = require('semver');
 const dockerApi = require('../../datasource/docker');
-const versions = require('../../workers/package/versions');
+const versions = require('../npm/versions');
 const compareVersions = require('compare-versions');
 
 module.exports = {
