+++ lib/manager/npm/package.js
@@ -1,5 +1,5 @@
 const npmApi = require('../../datasource/npm');
-const versions = require('../../workers/package/versions');
+const versions = require('./versions');
 const nodeManager = require('../_helpers/node/package');
 
 module.exports = {
