+++ lib/manager/npm/versions.js
